requirements.txt
asgiref==3.8.1
certifi==2024.2.2
distlib==0.3.8
Django==4.2.11
djangorestframework==3.15.1
filelock==3.14.0
pipenv==2023.12.1
platformdirs==4.2.1
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.11.0
virtualenv==20.26.1
